Fix an implicit 64-bit to 32-bit conversion.
../../third_party/crashpad/crashpad/util/misc/metrics.cc(66,35):  error: implicit conversion loses integer precision: 'crashpad::FileOffset' (aka 'long long') to 'base::HistogramBase::Sample' (aka 'int') [-Werror,-Wshorten-64-to-32]
      "Crashpad.CrashReportSize", size, 0, 20 * 1024 * 1024, 50);
      ~~~~~~~~~~~~~~~~~~~~~~~~~~~~^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

BUG=588506

+++ util/misc/metrics.cc
@@ -62,8 +62,11 @@ void Metrics::CrashReportPending(PendingReportReason reason) {
 
 // static
 void Metrics::CrashReportSize(FileOffset size) {
-  UMA_HISTOGRAM_CUSTOM_COUNTS(
-      "Crashpad.CrashReportSize", size, 0, 20 * 1024 * 1024, 50);
+  UMA_HISTOGRAM_CUSTOM_COUNTS("Crashpad.CrashReportSize",
+                              static_cast<int32_t>(size),
+                              0,
+                              20 * 1024 * 1024,
+                              50);
 }
 
 // static
